Add argsIgnorePattern for no-unused-vars
Changes this rule to allow unused arguments in functions, assuming that
the variable is prefixed with an underscore.

For example:

Under the current rules this is not permitted:

      const myMap = new Map([
        ['a', 1],
        ['b', 2],
        ['c', 3],
        ['d', 4],
      ]);

      Array.from(myMap, ([_key, value]) => value);
      > [1, 2, 3, 4]

or similarity

      Array.from(myMap, ([_, value]) => value);

And this is what would be preferred by the config

      Array.from(myMap, (entryPair) => entryPair[1]);

Oddly enough this is also allowed by the current config

      Array.from(myMap, ([, value]) => value);

Which is arguably much worse.

Ignoring unused positional arguments is quite common in functional
styles, which is what is encouraged.